backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.0
rest-framework-simplejwt==0.0.2
rpds-py==0.18.0
simplejwt==2.0.1
sqlparse==0.5.0
typing==3.7.4.3
typing_extensions==4.11.0
uritemplate==4.1.1
